relocate: update_simulation_validation_status to validation.py
This is the function which enables/disables users to run a sim
