+++ src/python/impactx/dashboard/Input/latticeConfiguration/latticeMain.py
@@ -76,7 +76,7 @@ def add_lattice_element():
     }
 
     state.selected_lattice_list.append(selected_lattice_element)
-    generalFunctions.update_simulation_validation_status()
+    DashboardValidation.update_simulation_validation_status()
     return selected_lattice_element
 
 
@@ -115,7 +115,7 @@ def parameter_input_checker_for_lattice(latticeElement):
 def on_selected_lattice_list_change(selected_lattice_list, **kwargs):
     if selected_lattice_list == []:
         state.isSelectedLatticeListEmpty = "Please select a lattice element"
-        generalFunctions.update_simulation_validation_status()
+        DashboardValidation.update_simulation_validation_status()
     else:
         state.isSelectedLatticeListEmpty = ""
 
@@ -207,7 +207,7 @@ def on_lattice_element_parameter_change(
             param["sim_input"] = sim_input
             param["parameter_error_message"] = error_message
 
-    generalFunctions.update_simulation_validation_status()
+    DashboardValidation.update_simulation_validation_status()
     state.dirty("selected_lattice_list")
