+++ src/python/impactx/dashboard/Input/space_charge_configuration/spaceChargeMain.py
@@ -108,31 +108,31 @@ def update_blocking_factor_and_n_cell(category, kwargs):
 def on_poisson_solver_change(poisson_solver, **kwargs):
     populate_prob_relative_fields()
     state.dirty("prob_relative_fields")
-    generalFunctions.update_simulation_validation_status()
+    DashboardValidation.update_simulation_validation_status()
 
 
 @state.change("space_charge")
 def on_space_charge_change(space_charge, **kwargs):
     state.dynamic_size = space_charge != "false"
-    generalFunctions.update_simulation_validation_status()
+    DashboardValidation.update_simulation_validation_status()
 
 
 @state.change("max_level")
 def on_max_level_change(max_level, **kwargs):
     populate_prob_relative_fields()
-    generalFunctions.update_simulation_validation_status()
+    DashboardValidation.update_simulation_validation_status()
 
 
 @state.change("blocking_factor_x", "blocking_factor_y", "blocking_factor_z")
 def on_blocking_factor_change(**kwargs):
     update_blocking_factor_and_n_cell("blocking_factor", kwargs)
-    generalFunctions.update_simulation_validation_status()
+    DashboardValidation.update_simulation_validation_status()
 
 
 @state.change("n_cell_x", "n_cell_y", "n_cell_z")
 def on_n_cell_change(**kwargs):
     update_blocking_factor_and_n_cell("n_cell", kwargs)
-    generalFunctions.update_simulation_validation_status()
+    DashboardValidation.update_simulation_validation_status()
 
 
 @ctrl.add("update_prob_relative")
@@ -164,7 +164,7 @@ def on_update_prob_relative_call(index, value):
         state.prob_relative_fields[index + 1]["error_message"] = next_error_message
 
     state.dirty("prob_relative_fields")
-    generalFunctions.update_simulation_validation_status()
+    DashboardValidation.update_simulation_validation_status()
 
 
 # -----------------------------------------------------------------------------
